Refactor and Removal of old codes - Created PasswordReminderService
