+++ src/components/Main.ui.js
@@ -57,7 +57,6 @@ import { registerDevice } from '../modules/Device/action'
 import { logoutRequest } from '../modules/Login/action.js'
 import ControlPanel from '../modules/UI/components/ControlPanel/ControlPanelConnector'
 import { ifLoggedIn } from '../modules/UI/components/LoginStatus/LoginStatus.js'
-import { passwordReminderModalConnector as PasswordReminderModal } from '../modules/UI/components/PasswordReminderModal/indexPasswordReminderModal.js'
 import { type Permission } from '../reducers/PermissionsReducer.js'
 import { type Dispatch, type RootState } from '../types/reduxTypes.js'
 import { scale } from '../util/scaling.js'
@@ -150,7 +149,6 @@ export class MainComponent extends React.Component<Props> {
             {this.renderTabView()}
           </Stack>
         </RouterWithRedux>
-        <PasswordReminderModal />
       </>
     )
   }
