--- src/modules/UI/components/PasswordReminderModal/indexPasswordReminderModal.js
@@ -1,31 +1,0 @@
-// @flow
-
-import {
-  checkPassword,
-  checkPasswordFail,
-  checkPasswordStart,
-  checkPasswordSuccess,
-  postponePasswordReminder,
-  requestChangePassword,
-  setPasswordReminder
-} from './actions.js'
-import { PasswordReminderModal } from './PasswordReminderModal.ui.js'
-import { passwordReminderModalConnector } from './passwordReminderModalConnector.js'
-import type { PasswordReminderModalState } from './passwordReminderModalReducer.js'
-import { initialState, passwordReminderModalReducer } from './passwordReminderModalReducer.js'
-
-export {
-  initialState,
-  checkPassword,
-  checkPasswordStart,
-  checkPasswordSuccess,
-  checkPasswordFail,
-  requestChangePassword,
-  postponePasswordReminder,
-  passwordReminderModalConnector,
-  PasswordReminderModal,
-  passwordReminderModalReducer as passwordReminderModal,
-  setPasswordReminder
-}
-
-export type { PasswordReminderModalState }
